Fixed the issues of including Thrust and MIOpen

+++ modules/module6/examples/01_convolution_hip.cpp
@@ -296,7 +296,7 @@ __device__ float wavefront_reduce_sum(float val) {
 void initialize_data(float *data, int size, bool random = true) {
     for (int i = 0; i < size; i++) {
         if (random) {
-            data[i] = static_cast<float>(rand()) / RAND_MAX;
+            data[i] = static_cast<float>(rand()) / static_cast<float>(RAND_MAX);
         } else {
             data[i] = 1.0f; // Unit impulse for testing
         }
